Extract run_plan tests to own file.
